+++ CHANGELOG.md
@@ -1,14 +1,19 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [1.4.0] - 2020-11-23
-- change bundle configuration
-- index exports `components` and `loadingBar`
+## [1.4.1] - 2020-11-24
+- remove delay from actual display of route and "routerload" event dispatch
+- change `loading-bar.claimed` to `loading-bar.claimedBy` (previous one still available)
+- add `loading-bar.isLoading`
 
 ## [any] - Unreleased
 - "requestvisibility" documentation
 
 - customable `z-index` of loading bar
+
+## [1.4.0] - 2020-11-23
+- change bundle configuration
+- index exports `components` and `loadingBar`
 
 ## [1.3.12] - 2020-11-18
 - unroutes correctly when loading routes overlap